+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/utils/internal/ClassCastsIncluded.kt
@@ -1,5 +1,13 @@
 package dev.inmo.tgbotapi.utils.internal
 
+import dev.inmo.tgbotapi.utils.RiskFeature
+
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.SOURCE)
-internal annotation class ClassCastsIncluded
+@RiskFeature("It is internal API in tgbotapi.core and should not be used outside")
+annotation class ClassCastsIncluded(val typesRegex: String = "", val excludeRegex: String = "")
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.SOURCE)
+@RiskFeature("It is internal API in tgbotapi.core and should not be used outside")
+annotation class ClassCastsExcluded