+++ tgbotapi.ksp/src/main/kotlin/TelegramBotAPISymbolProcessor.kt
@@ -1,12 +1,21 @@
 package dev.inmo.tgbotapi.ksp.processor
 
+import com.google.devtools.ksp.KspExperimental
 import com.google.devtools.ksp.getAllSuperTypes
+import com.google.devtools.ksp.getAnnotationsByType
+import com.google.devtools.ksp.isAnnotationPresent
 import com.google.devtools.ksp.processing.*
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.squareup.kotlinpoet.AnnotationSpec
+import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
+import com.squareup.kotlinpoet.asClassName
+import com.squareup.kotlinpoet.ksp.toClassName
 import com.squareup.kotlinpoet.ksp.writeTo
+import dev.inmo.tgbotapi.utils.RiskFeature
+import dev.inmo.tgbotapi.utils.internal.ClassCastsExcluded
+import dev.inmo.tgbotapi.utils.internal.ClassCastsIncluded
 import java.io.File
 
 class TelegramBotAPISymbolProcessor(
@@ -15,24 +24,51 @@ class TelegramBotAPISymbolProcessor(
     private val outputFile: String = "Output",
     private val outputFolder: String? = null
 ) : SymbolProcessor {
+    private val classCastsIncludedClassName = ClassCastsIncluded::class.asClassName()
+    @OptIn(KspExperimental::class, RiskFeature::class)
     override fun process(resolver: Resolver): List<KSAnnotated> {
-        val classes = resolver.getSymbolsWithAnnotation("dev.inmo.tgbotapi.utils.internal.ClassCastsIncluded").filterIsInstance<KSClassDeclaration>()
+        val classes = resolver.getSymbolsWithAnnotation(classCastsIncludedClassName.canonicalName).filterIsInstance<KSClassDeclaration>()
+        val classesRegexes: Map<KSClassDeclaration, Pair<Regex?, Regex?>> = classes.mapNotNull {
+            it to (it.getAnnotationsByType(ClassCastsIncluded::class).firstNotNullOfOrNull {
+                it.typesRegex.takeIf { it.isNotEmpty() } ?.let(::Regex) to it.excludeRegex.takeIf { it.isNotEmpty() } ?.let(::Regex)
+            } ?: return@mapNotNull null)
+        }.toMap()
         val classesSubtypes = mutableMapOf<KSClassDeclaration, MutableSet<KSClassDeclaration>>()
 
         resolver.getAllFiles().forEach {
             it.declarations.forEach { potentialSubtype ->
-                if (potentialSubtype is KSClassDeclaration) {
+                if (
+                    potentialSubtype is KSClassDeclaration
+                    && potentialSubtype.isAnnotationPresent(ClassCastsExcluded::class).not()
+                ) {
                     val allSupertypes = potentialSubtype.getAllSuperTypes().map { it.declaration }
-                    classes.forEach {
-                        if (it in allSupertypes) {
-                            classesSubtypes.getOrPut(it) { mutableSetOf() }.add(potentialSubtype)
+
+                    for (currentClass in classes) {
+                        val regexes = classesRegexes[currentClass]
+                        val simpleName = potentialSubtype.simpleName.getShortName()
+                        when {
+                            currentClass !in allSupertypes
+                            || regexes ?.first ?.matches(simpleName) == false
+                            || regexes ?.second ?.matches(simpleName) == true -> continue
+                            else -> {
+                                classesSubtypes.getOrPut(currentClass) { mutableSetOf() }.add(potentialSubtype)
+                            }
                         }
                     }
                 }
             }
         }
         fun fillWithSealeds(source: KSClassDeclaration, current: KSClassDeclaration = source) {
+            val regexes = classesRegexes[source]
             current.getSealedSubclasses().forEach {
+                val simpleName = it.simpleName.getShortName()
+                if (
+                    regexes ?.first ?.matches(simpleName) == false
+                    || regexes ?.second ?.matches(simpleName) == true
+                    || it.isAnnotationPresent(ClassCastsExcluded::class)
+                ) {
+                    return@forEach
+                }
                 classesSubtypes.getOrPut(source) { mutableSetOf() }.add(it)
                 fillWithSealeds(source, it)
             }