+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # TelegramBotAPI changelog
 
+## 8.1.0
+
+**PARTIALLY BREAKING CHANGES: Exclude `.*Impl` classcasts from `ClassCastsNew`**
+
+* `Version`:
+  * `MicroUtils`: `0.19.2` -> `0.19.4`
+* `Utils`:
+    * Add deep links formatting for internal `tg://` prefix (thanks to [@klimatov](https://github.com/klimatov))
+    * Exclude `.*Impl` classcasts from `ClassCastsNew`
+
 ## 8.0.1
 
 * `Version`: